requirements.txt
colorama~=0.4.5
PyYAML~=6.0
selenium~=4.5.0
python-dotenv~=0.21.0
requests~=2.28.1
chromedriver-autoinstaller==0.4.0
cryptography==38.0.1
